requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.3
cryptography==36.0.2
Flask==2.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
monotonic==1.6
neo4j==4.4.2
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pansi==2020.7.3
py2neo==2021.2.3
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
typing==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
